+++ scripts/group.py
@@ -15,7 +15,7 @@ group1 = viewer.scene.add([box1, box2])
 # with kwargs for each object
 box1 = box1.transformed(Translation.from_vector([0, 5, 0]))
 box2 = box2.transformed(Translation.from_vector([0, 5, 0]))
-group2 = viewer.scene.add([(box1, {"name": "box1", "surfacecolor": Color.red()}), (box2, {"name": "box2", "surfacecolor": Color.green()})])
+group2 = viewer.scene.add([(box1, {"name": "box1", "facecolor": Color.red()}), (box2, {"name": "box2", "facecolor": Color.green()})], linecolor=Color.blue())
 
 # with nested groups
 box1 = box1.transformed(Translation.from_vector([0, 5, 0]))